# src/utils/oas-spec.json
{
  "info": {
    "title": "Blue Prism AutomateC API",
    "version": "0.1.0",
    "description": "[Specification JSON](/api-spec)"
  },
  "paths": {
    "/version": {
      "get": {
        "summary": "Returns API version",
        "consumes": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "version": {
                  "type": "string",
                  "example": "0.1.0"
                }
              }
            }
          }
        },
        "description": "Returns version of currently running API server.",
        "produces": ["application/json"]
      }
    },
    "/processes": {
      "post": {
        "summary": "Starts a new process",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "process",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "example": {
                "process": "Login"
              },
              "properties": {
                "process": "string"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "sessionId": {
                  "type": "string",
                  "example": "6ee248d1-72f3-4890-af5c-ca9ea6ce4214"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "The maximum number of concurrent sessions permitted by the current license would be exceeded"
                }
              }
            }
          }
        },
        "description": "Requests starting new process. Process name must be supplied",
        "produces": ["application/json"]
      }
    },
    "/processes/{sessionId}": {
      "get": {
        "summary": "Returns status of the process",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "example": "6ee248d1-72f3-4890-af5c-ca9ea6ce4214"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Running"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No information found for that session"
                }
              }
            }
          }
        },
        "description": "Returns status of the process supplied in the URL path. Session ID (usually obtained when starting the process) must be supplied.",
        "produces": ["application/json"]
      }
    },
    "/processes/{sessionId}/stop": {
      "post": {
        "summary": "Requests process to stop",
        "consumes": ["application/json"],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "example": "6ee248d1-72f3-4890-af5c-ca9ea6ce4214"
            }
          }
        ],
        "responses": {
          "202": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "Stop requested"
                }
              }
            }
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Could not find the session with the ID/number"
                }
              }
            }
          }
        },
        "description": "Requests  process supplied in the URL path to stop. Session ID (usually obtained when starting the process) must be supplied.",
        "produces": ["application/json"]
      }
    }
  },
  "security": {
    "UserSecurity": []
  },
  "securityDefinition": {
    "UserSecurity": {
      "type": "basic"
    }
  },
  "swagger": "2.0"
}